feat: initial commit of android support
Precommit-Verified: d694e74547584dc155fdf395def59eadb4157b1efef9ad43e62803dea29c07d7

+++ src/lhashes_sha1.cpp
@@ -191,7 +191,8 @@ void SHA1::update (const unsigned char* data, std::size_t length)
 	{
 		char sbuf[BLOCK_BYTES] = {};
 
-		const auto count = std::min (length, BLOCK_BYTES - buffer.size());
+		const auto count = std::min (static_cast<size_t>(length),
+									 static_cast<size_t>(BLOCK_BYTES - buffer.size()));
 
 		std::memcpy (sbuf, data, count);
